third_party/WebKit/LayoutTests/inspector-protocol/heap-profiler/sampling-heap-profiler.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

function testMain()
{
    makeDeepCallStack(10, junkGenerator);
}

function makeDeepCallStack(depth, action)
{
    if (depth)
        makeDeepCallStack(depth - 1, action);
    else
        action();
}

function junkGenerator()
{
    var junkArray = new Array(3000);
    window.junkArray = junkArray;
    for (var i = 1; i < junkArray.length; ++i)
        junkArray[i] = new Array(i);
}

function test()
{
    InspectorTest.sendCommand("HeapProfiler.startSampling", {}, samplingStarted);

    function samplingStarted()
    {
        InspectorTest.log("Sampling started");
        InspectorTest.sendCommand("Runtime.evaluate", { expression: "testMain()" }, stopSampling);
    }

    function stopSampling()
    {
        InspectorTest.sendCommand("HeapProfiler.stopSampling", {}, samplingStopped);
    }

    function samplingStopped(message)
    {
        InspectorTest.log("Sampling stopped");
        var profile = message.result.profile;
        var head = profile.head;
        logNode(findNode(head, "testMain", 1));
        logNode(findNode(head, "makeDeepCallStack", 11));
        logNode(findNode(head, "junkGenerator", 12));

        InspectorTest.completeTest();
    }

    function findNode(root, name, depth)
    {
        if (depth < 1 && root.callFrame.functionName === name)
            return root;
        return root.children.reduce((found, child) => found || findNode(child, name, depth - 1), null);
    }

    function logNode(node)
    {
        var size = typeof node.selfSize === "number" ? node.selfSize ? ">0" : "=0" : "-";
        InspectorTest.log(`size${size}   ${node.callFrame.functionName}:${node.callFrame.lineNumber}:${node.callFrame.columnNumber}`);
    }

    //# sourceURL=sampling-heap-profiler.html
}
</script>
</head>
<body onload="runTest()">
<p>Test sampling heap profiler.</p>
</body>
</html>
