third_party/WebKit/LayoutTests/inspector-protocol/heap-profiler/sampling-heap-profiler-expected.txt

Test sampling heap profiler.

Sampling started
Sampling stopped
size=0   testMain:11:18
size=0   makeDeepCallStack:16:27
size>0   junkGenerator:24:23
